Пересобрал классы

+++ src/main/java/ru/tigran/player/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package ru.tigran.player.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import ru.tigran.player.model.UserEntity;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByUsername(String username);
-} 
+}
