+++ src/main/java/ru/tigran/player/repository/VideoRepository.java
@@ -3,5 +3,5 @@ package ru.tigran.player.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.tigran.player.model.VideoEntity;
 
-public interface VideoRepository extends JpaRepository<VideoEntity, Integer> {
+public interface VideoRepository extends JpaRepository<VideoEntity, Long> {
 }
